refactor(scim): split users.go into per-RPC files (#149 slice 4)
Fourth slice. Splits users.go's per-RPC handlers across three sibling
files so users.go drops under the issue's <500 LOC bar.

- users_create.go: createUser (the four-outcome handler with the
  AutoLinkByEmail + idempotent-POST + auto-enable-provisioning
  branches)
- users_mutate.go: replaceUser, patchUser, handleUserPatchReplace
  (the SCIM PUT + PATCH dispatchers)
- users.go: listUsers, listUsersFiltered, getUser, deleteUser only

Size delta:
- users.go: 897 \xe2\x86\x92 304 LOC (-593) \xe2\x86\x90 ACCEPTANCE BAR REACHED
- New: users_create.go (254), users_mutate.go (385)
- Cumulative since baseline 1273 LOC: -969 (-76%)

users.go now reads at a glance as the four primitive HTTP routes
(GET list, GET one, GET filtered, DELETE one) plus the listFiltered
dispatcher. The mutating handlers are isolated where their
complexity belongs.

Pure file-shuffling \xe2\x80\x94 zero behaviour change. Stacked on #232
(slice 2 users_helpers); will rebase as it lands.

Refs manchtools/power-manage-server#149 (audit F009).
