+++ internal/scim/users.go
@@ -1,13 +1,10 @@
 package scim
 
 import (
-	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"net/http"
 	"strconv"
-	"strings"
 
 	"github.com/jackc/pgx/v5"
 
@@ -153,232 +150,6 @@ func (h *Handler) listUsersFiltered(w http.ResponseWriter, r *http.Request, prov
 }
 
 // createUser handles POST /scim/v2/{slug}/Users
-func (h *Handler) createUser(w http.ResponseWriter, r *http.Request) {
-	h.logger.Debug("SCIM createUser called")
-	provider, ok := providerFromContext(r.Context())
-	if !ok {
-		writeError(w, http.StatusUnauthorized, "not authenticated")
-		return
-	}
-
-	var scimUser SCIMUser
-	limitBody(r)
-	if err := json.NewDecoder(r.Body).Decode(&scimUser); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid JSON body")
-		return
-	}
-
-	// Extract email: prefer userName, fall back to emails[0].value
-	email := scimUser.UserName
-	if email == "" && len(scimUser.Emails) > 0 {
-		email = scimUser.Emails[0].Value
-	}
-	if email == "" {
-		writeError(w, http.StatusBadRequest, "userName or emails[0].value is required")
-		return
-	}
-
-	externalID := scimUser.ExternalID
-
-	ctx := r.Context()
-	baseURL := baseURLFromRequest(r, provider.Slug)
-
-	// Check if a user with this external ID already exists for this provider
-	if externalID != "" {
-		existing, err := h.store.Queries().FindSCIMUserByExternalID(ctx, db.FindSCIMUserByExternalIDParams{
-			ProviderID: provider.ID,
-			ExternalID: externalID,
-		})
-		if err == nil {
-			// User already linked with this external ID — sync data from SCIM
-			h.logger.Debug("SCIM createUser: user already exists by external ID, syncing", "user_id", existing.ID, "external_id", externalID)
-			// (source of truth) and return existing resource. Using 200 instead
-			// of 409 makes POST idempotent for SCIM clients that re-POST on
-			// every sync cycle.
-			h.syncUserFromSCIM(ctx, provider, existing.ID, email, scimUser.Active, scimUser.Name)
-			user, err := h.store.Queries().GetUserByID(ctx, existing.ID)
-			if err != nil {
-				writeJSON(w, http.StatusOK, findExternalIDUserRowToSCIM(existing, baseURL))
-			} else {
-				writeJSON(w, http.StatusOK, userToSCIM(user, existing.ScimExternalID, baseURL))
-			}
-			return
-		}
-		if !errors.Is(err, pgx.ErrNoRows) {
-			h.logger.Error("failed to check existing SCIM user", "error", err)
-			writeError(w, http.StatusInternalServerError, "internal server error")
-			return
-		}
-	}
-
-	// Check if we should link to an existing user by email
-	if provider.AutoLinkByEmail {
-		existing, err := h.store.Queries().FindSCIMUserByEmail(ctx, db.FindSCIMUserByEmailParams{
-			ProviderID: provider.ID,
-			Email:      email,
-		})
-		if err == nil {
-			// Already linked — sync data from SCIM (source of truth) and return
-			h.syncUserFromSCIM(ctx, provider, existing.ID, email, scimUser.Active, scimUser.Name)
-			user, readErr := h.store.Queries().GetUserByID(ctx, existing.ID)
-			if readErr != nil {
-				writeJSON(w, http.StatusOK, findUserRowToSCIM(existing, baseURL))
-			} else {
-				writeJSON(w, http.StatusOK, userToSCIM(user, existing.ScimExternalID, baseURL))
-			}
-			return
-		}
-
-		// Check if user exists but is not yet linked
-		existingUser, err := h.store.Queries().GetUserByEmail(ctx, email)
-		if err == nil {
-			// User exists — create identity link
-			h.logger.Debug("SCIM createUser: linking existing user by email", "user_id", existingUser.ID, "email", email)
-			linkID := newULID()
-			err = h.store.AppendEvent(ctx, store.Event{
-				StreamType: "identity_provider",
-				StreamID:   linkID,
-				EventType:  string(eventtypes.IdentityLinked),
-				Data: map[string]any{
-					"user_id":        existingUser.ID,
-					"provider_id":    provider.ID,
-					"external_id":    externalID,
-					"external_email": email,
-					"external_name":  formatExternalName(scimUser.Name),
-				},
-				ActorType: "scim",
-				ActorID:   provider.ID,
-			})
-			if err != nil {
-				h.logger.Error("failed to link existing user via SCIM", "error", err)
-				writeError(w, http.StatusInternalServerError, "failed to link user")
-				return
-			}
-
-			writeJSON(w, http.StatusCreated, userToSCIM(existingUser, externalID, baseURL))
-			return
-		}
-		if !errors.Is(err, pgx.ErrNoRows) {
-			h.logger.Error("failed to look up user by email", "error", err)
-			writeError(w, http.StatusInternalServerError, "internal server error")
-			return
-		}
-	}
-
-	// Create new user
-	h.logger.Debug("SCIM createUser: creating new user", "email", email, "external_id", externalID)
-	userID := newULID()
-
-	linuxUID, err := h.store.Queries().GetNextLinuxUID(ctx)
-	if err != nil {
-		h.logger.Error("failed to assign linux uid via SCIM", "error", err)
-		writeError(w, http.StatusInternalServerError, "failed to assign linux uid")
-		return
-	}
-	linuxUsername := deriveLinuxUsername(email, scimUser.UserName)
-	if linuxUsername == "" {
-		linuxUsername = "user_" + userID[:8]
-	}
-
-	// Resolve the role ID set BEFORE emitting the event so the user
-	// INSERT and the per-role INSERT land atomically inside the
-	// projector's WithTx (issue #135). SCIM only ever assigns the
-	// provider's configured default role at creation time; if no
-	// default is configured the slice stays empty and the projector
-	// skips the per-role INSERT loop.
-	var roleIDs []string
-	if provider.DefaultRoleID != "" {
-		roleIDs = []string{provider.DefaultRoleID}
-	}
-
-	err = h.store.AppendEvent(ctx, store.Event{
-		StreamType: "user",
-		StreamID:   userID,
-		EventType:  string(eventtypes.UserCreatedWithRoles),
-		Data: map[string]any{
-			"email":          email,
-			"display_name":   formatExternalName(scimUser.Name),
-			"given_name":     safeNameField(scimUser.Name, "given"),
-			"family_name":    safeNameField(scimUser.Name, "family"),
-			"linux_username": linuxUsername,
-			"linux_uid":      linuxUID,
-			"role_ids":       roleIDs,
-		},
-		ActorType: "scim",
-		ActorID:   provider.ID,
-	})
-	if err != nil {
-		h.logger.Error("failed to create user via SCIM", "error", err)
-		writeError(w, http.StatusInternalServerError, "failed to create user")
-		return
-	}
-
-	// Create identity link
-	linkID := newULID()
-	err = h.store.AppendEvent(ctx, store.Event{
-		StreamType: "identity_provider",
-		StreamID:   linkID,
-		EventType:  string(eventtypes.IdentityLinked),
-		Data: map[string]any{
-			"user_id":        userID,
-			"provider_id":    provider.ID,
-			"external_id":    externalID,
-			"external_email": email,
-			"external_name":  formatExternalName(scimUser.Name),
-		},
-		ActorType: "scim",
-		ActorID:   provider.ID,
-	})
-	if err != nil {
-		h.logger.Error("failed to create identity link via SCIM", "error", err)
-		writeError(w, http.StatusInternalServerError, "failed to link user")
-		return
-	}
-
-	// Auto-enable provisioning/SSH if global server settings are on
-	if settings, err := h.store.Queries().GetServerSettings(ctx); err == nil {
-		if settings.UserProvisioningEnabled {
-			if err := h.store.AppendEvent(ctx, store.Event{
-				StreamType: "user",
-				StreamID:   userID,
-				EventType:  string(eventtypes.UserProvisioningSettingsUpdated),
-				Data:       map[string]any{"user_provisioning_enabled": true},
-				ActorType:  "system",
-				ActorID:    "scim",
-			}); err != nil {
-				h.logger.Warn("failed to auto-enable provisioning for SCIM user", "user_id", userID, "error", err)
-			}
-		}
-		if settings.SshAccessForAll {
-			if err := h.store.AppendEvent(ctx, store.Event{
-				StreamType: "user",
-				StreamID:   userID,
-				EventType:  string(eventtypes.UserSshSettingsUpdated),
-				Data: map[string]any{
-					"ssh_access_enabled": true,
-					"ssh_allow_pubkey":   true,
-					"ssh_allow_password": false,
-				},
-				ActorType: "system",
-				ActorID:   "scim",
-			}); err != nil {
-				h.logger.Warn("failed to auto-enable SSH for SCIM user", "user_id", userID, "error", err)
-			}
-		}
-	} else {
-		h.logger.Warn("failed to check server settings for SCIM user defaults", "error", err)
-	}
-
-	// Read back created user
-	user, err := h.store.Queries().GetUserByID(ctx, userID)
-	if err != nil {
-		h.logger.Error("failed to read back created user", "error", err)
-		writeError(w, http.StatusInternalServerError, "user created but failed to read back")
-		return
-	}
-
-	writeJSON(w, http.StatusCreated, userToSCIM(user, externalID, baseURL))
-}
 
 // getUser handles GET /scim/v2/{slug}/Users/{id}
 func (h *Handler) getUser(w http.ResponseWriter, r *http.Request) {
@@ -429,370 +200,6 @@ func (h *Handler) getUser(w http.ResponseWriter, r *http.Request) {
 }
 
 // replaceUser handles PUT /scim/v2/{slug}/Users/{id}
-func (h *Handler) replaceUser(w http.ResponseWriter, r *http.Request) {
-	h.logger.Debug("SCIM replaceUser called")
-	provider, ok := providerFromContext(r.Context())
-	if !ok {
-		writeError(w, http.StatusUnauthorized, "not authenticated")
-		return
-	}
-
-	userID := r.PathValue("id")
-	if userID == "" {
-		writeError(w, http.StatusBadRequest, "missing user id")
-		return
-	}
-
-	// Verify this provider owns the user
-	if err := h.verifyProviderOwnership(r.Context(), provider.ID, userID); err != nil {
-		writeError(w, http.StatusNotFound, "user not found")
-		return
-	}
-
-	var scimUser SCIMUser
-	limitBody(r)
-	if err := json.NewDecoder(r.Body).Decode(&scimUser); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid JSON body")
-		return
-	}
-
-	ctx := r.Context()
-	baseURL := baseURLFromRequest(r, provider.Slug)
-
-	// Verify user exists
-	existingUser, err := h.store.Queries().GetUserByID(ctx, userID)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			writeError(w, http.StatusNotFound, "user not found")
-			return
-		}
-		h.logger.Error("failed to get user for replace", "error", err)
-		writeError(w, http.StatusInternalServerError, "failed to get user")
-		return
-	}
-
-	// Update email if changed
-	newEmail := scimUser.UserName
-	if newEmail == "" && len(scimUser.Emails) > 0 {
-		newEmail = scimUser.Emails[0].Value
-	}
-	if newEmail != "" && newEmail != existingUser.Email {
-		err = h.store.AppendEvent(ctx, store.Event{
-			StreamType: "user",
-			StreamID:   userID,
-			EventType:  string(eventtypes.UserEmailChanged),
-			Data: map[string]any{
-				"email": newEmail,
-			},
-			ActorType: "scim",
-			ActorID:   provider.ID,
-		})
-		if err != nil {
-			h.logger.Error("failed to update user email", "error", err)
-			writeError(w, http.StatusInternalServerError, "failed to update user email")
-			return
-		}
-	}
-
-	// Update active status
-	if !scimUser.IsActive() && !existingUser.Disabled {
-		if err := h.store.AppendEvent(ctx, store.Event{
-			StreamType: "user",
-			StreamID:   userID,
-			EventType:  string(eventtypes.UserDisabled),
-			Data:       map[string]any{},
-			ActorType:  "scim",
-			ActorID:    provider.ID,
-		}); err != nil {
-			h.logger.Error("failed to disable user via SCIM", "error", err)
-			writeError(w, http.StatusInternalServerError, "failed to update user status")
-			return
-		}
-	} else if scimUser.IsActive() && existingUser.Disabled {
-		if err := h.store.AppendEvent(ctx, store.Event{
-			StreamType: "user",
-			StreamID:   userID,
-			EventType:  string(eventtypes.UserEnabled),
-			Data:       map[string]any{},
-			ActorType:  "scim",
-			ActorID:    provider.ID,
-		}); err != nil {
-			h.logger.Error("failed to enable user via SCIM", "error", err)
-			writeError(w, http.StatusInternalServerError, "failed to update user status")
-			return
-		}
-	}
-
-	// Update profile if name fields provided
-	newDisplayName := formatExternalName(scimUser.Name)
-	newGivenName := safeNameField(scimUser.Name, "given")
-	newFamilyName := safeNameField(scimUser.Name, "family")
-	if newDisplayName != "" || newGivenName != "" || newFamilyName != "" {
-		if err := h.store.AppendEvent(ctx, store.Event{
-			StreamType: "user",
-			StreamID:   userID,
-			EventType:  string(eventtypes.UserProfileUpdated),
-			Data: map[string]any{
-				"display_name": newDisplayName,
-				"given_name":   newGivenName,
-				"family_name":  newFamilyName,
-			},
-			ActorType: "scim",
-			ActorID:   provider.ID,
-		}); err != nil {
-			h.logger.Warn("failed to update user profile via SCIM", "error", err)
-		}
-	}
-
-	// Sync identity link (external_email + external_name) from SCIM source of truth
-	effectiveEmail := newEmail
-	if effectiveEmail == "" {
-		effectiveEmail = existingUser.Email
-	}
-	h.syncIdentityLink(ctx, provider, userID, effectiveEmail, scimUser.Name)
-
-	// Read back updated user
-	user, err := h.store.Queries().GetUserByID(ctx, userID)
-	if err != nil {
-		h.logger.Error("failed to read back updated user", "error", err)
-		writeError(w, http.StatusInternalServerError, "failed to read user")
-		return
-	}
-
-	externalID := scimUser.ExternalID
-	if externalID == "" {
-		// Try to look up existing external ID
-		linked, linkErr := h.store.Queries().FindSCIMUserByEmail(ctx, db.FindSCIMUserByEmailParams{
-			ProviderID: provider.ID,
-			Email:      user.Email,
-		})
-		if linkErr == nil {
-			externalID = linked.ScimExternalID
-		}
-	}
-
-	writeJSON(w, http.StatusOK, userToSCIM(user, externalID, baseURL))
-}
-
-// patchUser handles PATCH /scim/v2/{slug}/Users/{id}
-func (h *Handler) patchUser(w http.ResponseWriter, r *http.Request) {
-	h.logger.Debug("SCIM patchUser called")
-	provider, ok := providerFromContext(r.Context())
-	if !ok {
-		writeError(w, http.StatusUnauthorized, "not authenticated")
-		return
-	}
-
-	userID := r.PathValue("id")
-	if userID == "" {
-		writeError(w, http.StatusBadRequest, "missing user id")
-		return
-	}
-
-	// Verify this provider owns the user
-	if err := h.verifyProviderOwnership(r.Context(), provider.ID, userID); err != nil {
-		writeError(w, http.StatusNotFound, "user not found")
-		return
-	}
-
-	limitBody(r)
-	var patch SCIMPatchRequest
-	if err := json.NewDecoder(r.Body).Decode(&patch); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid JSON body")
-		return
-	}
-
-	ctx := r.Context()
-	baseURL := baseURLFromRequest(r, provider.Slug)
-
-	// Verify user exists
-	existingUser, err := h.store.Queries().GetUserByID(ctx, userID)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			writeError(w, http.StatusNotFound, "user not found")
-			return
-		}
-		h.logger.Error("failed to get user for patch", "error", err)
-		writeError(w, http.StatusInternalServerError, "failed to get user")
-		return
-	}
-
-	for _, op := range patch.Operations {
-		// Pre-validate per RFC 7644 §3.5.2 so the user patch path
-		// rejects unknown verbs with a 400 even though add/remove
-		// fall through to the catch-all default below — they are
-		// not implemented for users today, but they ARE valid SCIM
-		// ops, and a 400 with the precise error is the right
-		// response (vs. a 501 for the implementation gap).
-		if !op.Op.IsValid() {
-			writeError(w, http.StatusBadRequest, fmt.Sprintf("unsupported patch op: %s", op.Op))
-			return
-		}
-		switch op.Op.Normalize() {
-		case SCIMPatchOpReplace:
-			if err := h.handleUserPatchReplace(ctx, provider, userID, existingUser, op); err != nil {
-				h.logger.Error("failed to apply SCIM patch op", "op", op.Op, "path", op.Path, "error", err)
-				writeError(w, http.StatusInternalServerError, "failed to apply patch operation")
-				return
-			}
-		default:
-			writeError(w, http.StatusBadRequest, fmt.Sprintf("unsupported patch op: %s", op.Op))
-			return
-		}
-	}
-
-	// Sync identity link with any name/email changes from PATCH ops
-	patchedUser, err := h.store.Queries().GetUserByID(ctx, userID)
-	if err != nil {
-		h.logger.Error("failed to read back patched user", "error", err)
-		writeError(w, http.StatusInternalServerError, "failed to read user")
-		return
-	}
-	h.syncIdentityLink(ctx, provider, userID, patchedUser.Email, extractNameFromPatchOps(patch.Operations))
-
-	// Look up external ID
-	externalID := ""
-	linked, linkErr := h.store.Queries().FindSCIMUserByEmail(ctx, db.FindSCIMUserByEmailParams{
-		ProviderID: provider.ID,
-		Email:      patchedUser.Email,
-	})
-	if linkErr == nil {
-		externalID = linked.ScimExternalID
-	}
-
-	writeJSON(w, http.StatusOK, userToSCIM(patchedUser, externalID, baseURL))
-}
-
-// handleUserPatchReplace processes a single "replace" patch operation on a user.
-func (h *Handler) handleUserPatchReplace(ctx context.Context, provider db.IdentityProvidersProjection, userID string, existingUser db.UsersProjection, op SCIMPatchOp) error {
-	path := strings.ToLower(op.Path)
-
-	switch path {
-	case "active":
-		active, ok := op.Value.(bool)
-		if !ok {
-			// Try string conversion
-			if s, ok := op.Value.(string); ok {
-				active = strings.EqualFold(s, "true")
-			}
-		}
-		if !active && !existingUser.Disabled {
-			return h.store.AppendEvent(ctx, store.Event{
-				StreamType: "user",
-				StreamID:   userID,
-				EventType:  string(eventtypes.UserDisabled),
-				Data:       map[string]any{},
-				ActorType:  "scim",
-				ActorID:    provider.ID,
-			})
-		} else if active && existingUser.Disabled {
-			return h.store.AppendEvent(ctx, store.Event{
-				StreamType: "user",
-				StreamID:   userID,
-				EventType:  string(eventtypes.UserEnabled),
-				Data:       map[string]any{},
-				ActorType:  "scim",
-				ActorID:    provider.ID,
-			})
-		}
-
-	case "username":
-		email, ok := op.Value.(string)
-		if !ok || email == "" {
-			return fmt.Errorf("invalid userName value")
-		}
-		if email != existingUser.Email {
-			return h.store.AppendEvent(ctx, store.Event{
-				StreamType: "user",
-				StreamID:   userID,
-				EventType:  string(eventtypes.UserEmailChanged),
-				Data: map[string]any{
-					"email": email,
-				},
-				ActorType: "scim",
-				ActorID:   provider.ID,
-			})
-		}
-
-	case "emails":
-		// Value should be an array of email objects
-		emailsRaw, ok := op.Value.([]any)
-		if !ok || len(emailsRaw) == 0 {
-			return fmt.Errorf("invalid emails value")
-		}
-		emailObj, ok := emailsRaw[0].(map[string]any)
-		if !ok {
-			return fmt.Errorf("invalid email object")
-		}
-		email, ok := emailObj["value"].(string)
-		if !ok || email == "" {
-			return fmt.Errorf("invalid email value")
-		}
-		if email != existingUser.Email {
-			return h.store.AppendEvent(ctx, store.Event{
-				StreamType: "user",
-				StreamID:   userID,
-				EventType:  string(eventtypes.UserEmailChanged),
-				Data: map[string]any{
-					"email": email,
-				},
-				ActorType: "scim",
-				ActorID:   provider.ID,
-			})
-		}
-
-	case "name":
-		// Update profile fields from name object
-		nameMap, ok := op.Value.(map[string]any)
-		if !ok {
-			return fmt.Errorf("invalid name value")
-		}
-		data := map[string]any{}
-		if gn, ok := nameMap["givenName"].(string); ok {
-			data["given_name"] = gn
-		}
-		if fn, ok := nameMap["familyName"].(string); ok {
-			data["family_name"] = fn
-		}
-		if fm, ok := nameMap["formatted"].(string); ok {
-			data["display_name"] = fm
-		}
-		if len(data) > 0 {
-			return h.store.AppendEvent(ctx, store.Event{
-				StreamType: "user",
-				StreamID:   userID,
-				EventType:  string(eventtypes.UserProfileUpdated),
-				Data:       data,
-				ActorType:  "scim",
-				ActorID:    provider.ID,
-			})
-		}
-		// Identity link sync happens in patchUser after all ops.
-
-	case "name.givenname", "name.familyname", "name.formatted":
-		// Sub-path name changes — handled via identity link sync in patchUser.
-
-	case "":
-		// No path — the value is a map of attributes to replace
-		valueMap, ok := op.Value.(map[string]any)
-		if !ok {
-			return fmt.Errorf("replace without path requires object value")
-		}
-		for key, val := range valueMap {
-			subOp := SCIMPatchOp{
-				Op:    SCIMPatchOpReplace,
-				Path:  key,
-				Value: val,
-			}
-			if err := h.handleUserPatchReplace(ctx, provider, userID, existingUser, subOp); err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
-}
-
 // deleteUser handles DELETE /scim/v2/{slug}/Users/{id}
 func (h *Handler) deleteUser(w http.ResponseWriter, r *http.Request) {
 	h.logger.Debug("SCIM deleteUser called")
